fix: crash-proof reading-stats recording via periodic in-session flush
Reading stats were written only in the readers' onExit(), so any exit
path that never runs it -- a hang or watchdog reset on the sleep
transition, a battery pull, a crash -- silently lost the entire session.
Reported on X4: no day registered since Jul 13 and the streak broken,
while page progress (saved on page turns) and the clock stash (written
before onExit) kept working, which is exactly the signature of the exit
path being skipped.

ReaderUtils::flushReadingStats() now flushes whole elapsed minutes from
the reader's loop() every 5 minutes (one SD write per flush, matching
the write-throttling rule) and carries the sub-minute remainder forward;
onExit() keeps a force-flush for the tail. All four readers (EPUB,
manga, XTC, TXT) share the helper, replacing the duplicated inline
blocks. Stats now survive any exit path.
