+++ src/activities/reader/EpubReaderActivity.cpp
@@ -25,7 +25,6 @@
 #include <limits>
 
 #include "BookmarkEntry.h"
-#include <HalClock.h>
 
 #include "CrossPointSettings.h"
 #include "CrossPointState.h"
@@ -313,22 +312,9 @@ void EpubReaderActivity::onEnter() {
 void EpubReaderActivity::onExit() {
   Activity::onExit();
 
-  // Record reading time for stats
-  if (readingSessionStartMs > 0) {
-    unsigned long elapsed = millis() - readingSessionStartMs;
-    uint16_t minutes = static_cast<uint16_t>(elapsed / 60000);
-    if (minutes > 0) {
-      // Local-midnight day boundary: shift by the user's display UTC offset so an evening
-      // session doesn't get logged against "tomorrow" (UTC midnight is 9am in Japan).
-      time_t now = HalClock::localEpoch(SETTINGS.clockUtcOffsetQ);
-      struct tm* t = gmtime(&now);
-      READING_STATS.loadFromFile();
-      READING_STATS.addMinutes(static_cast<uint16_t>(t->tm_year + 1900),
-                               static_cast<uint8_t>(t->tm_mon + 1),
-                               static_cast<uint8_t>(t->tm_mday), minutes);
-      READING_STATS.saveToFile();
-    }
-  }
+  // Record the sub-interval tail of the session; whole minutes were already flushed
+  // periodically from loop() (see ReaderUtils::flushReadingStats).
+  ReaderUtils::flushReadingStats(readingSessionStartMs, /*force=*/true);
 
   // Reset orientation back to portrait for the rest of the UI
   renderer.setOrientation(GfxRenderer::Orientation::Portrait);
@@ -396,6 +382,10 @@ void EpubReaderActivity::loop() {
     finish();
     return;
   }
+
+  // Crash-proof stats: flush whole minutes every few minutes so an exit path that
+  // never reaches onExit() (hang/reset on sleep, battery pull) can't lose the day.
+  ReaderUtils::flushReadingStats(readingSessionStartMs);
 
   // End-of-Book screen reached (currentSpineIndex == spine count) means the book is
   // finished. Two independent finished-book features key off this same condition.
