+++ src/activities/reader/TxtReaderActivity.cpp
@@ -4,20 +4,16 @@
 #include <BidiUtils.h>
 #include <FontCacheManager.h>
 #include <GfxRenderer.h>
-#include <HalClock.h>
 #include <HalStorage.h>
 #include <I18n.h>
 #include <Serialization.h>
 #include <Utf8.h>
-
-#include <ctime>
 
 #include "CrossPointSettings.h"
 #include "CrossPointState.h"
 #include "MappedInputManager.h"
 #include "ProgressFile.h"
 #include "ReaderUtils.h"
-#include "ReadingStatsStore.h"
 #include "RecentBooksStore.h"
 #include "components/UITheme.h"
 #include "fontIds.h"
@@ -56,22 +52,10 @@ void TxtReaderActivity::onEnter() {
 void TxtReaderActivity::onExit() {
   Activity::onExit();
 
-  // Record reading time for stats (mirrors EpubReaderActivity::onExit). TXT books never
-  // counted toward the reading streak.
-  if (readingSessionStartMs > 0) {
-    const unsigned long elapsed = millis() - readingSessionStartMs;
-    const uint16_t minutes = static_cast<uint16_t>(elapsed / 60000);
-    if (minutes > 0) {
-      // Local-midnight day boundary: shift by the user's display UTC offset so an evening
-      // session doesn't get logged against "tomorrow" (UTC midnight is 9am in Japan).
-      const time_t now = HalClock::localEpoch(SETTINGS.clockUtcOffsetQ);
-      struct tm* t = gmtime(&now);
-      READING_STATS.loadFromFile();
-      READING_STATS.addMinutes(static_cast<uint16_t>(t->tm_year + 1900), static_cast<uint8_t>(t->tm_mon + 1),
-                               static_cast<uint8_t>(t->tm_mday), minutes);
-      READING_STATS.saveToFile();
-    }
-  }
+  // Record the sub-interval tail of the session; whole minutes were already flushed
+  // periodically from loop(). TXT books never counted toward the reading streak at all
+  // before this.
+  ReaderUtils::flushReadingStats(readingSessionStartMs, /*force=*/true);
 
   // Reset orientation back to portrait for the rest of the UI
   renderer.setOrientation(GfxRenderer::Orientation::Portrait);
@@ -84,6 +68,10 @@ void TxtReaderActivity::onExit() {
 }
 
 void TxtReaderActivity::loop() {
+  // Crash-proof stats: flush whole minutes every few minutes so an exit path that
+  // never reaches onExit() (hang/reset on sleep, battery pull) can't lose the day.
+  ReaderUtils::flushReadingStats(readingSessionStartMs);
+
   // Long press BACK (1s+) goes to file selection
   if (mappedInput.isPressed(MappedInputManager::Button::Back) && mappedInput.getHeldTime() >= ReaderUtils::GO_HOME_MS) {
     activityManager.goToFileBrowser(txt ? txt->getPath() : "");
